make a dedicated workflow to compute LUT, and only have LUT from file for main workflow
