+++ src/ess/reduce/time_of_flight/workflow.py
@@ -1,22 +1,25 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
 from collections.abc import Iterable
-from enum import Enum, auto
 
+# from enum import Enum, auto
 import sciline
 import scipp as sc
 
 from ..nexus import GenericNeXusWorkflow
-from . import eto_to_tof, simulation
-from .types import TimeOfFlightLookupTable, TimeOfFlightLookupTableFilename
+from . import eto_to_tof  # , simulation
+from .types import (
+    PulseStrideOffset,
+    TimeOfFlightLookupTable,
+    TimeOfFlightLookupTableFilename,
+)
 
+# class TofLutProvider(Enum):
+#     """Provider for the time-of-flight lookup table."""
 
-class TofLutProvider(Enum):
-    """Provider for the time-of-flight lookup table."""
-
-    FILE = auto()  # From file
-    TOF = auto()  # Computed with 'tof' package from chopper settings
-    MCSTAS = auto()  # McStas simulation (not implemented yet)
+#     FILE = auto()  # From file
+#     TOF = auto()  # Computed with 'tof' package from chopper settings
+#     MCSTAS = auto()  # McStas simulation (not implemented yet)
 
 
 def load_tof_lookup_table(
@@ -29,7 +32,7 @@ def GenericTofWorkflow(
     *,
     run_types: Iterable[sciline.typing.Key],
     monitor_types: Iterable[sciline.typing.Key],
-    tof_lut_provider: TofLutProvider = TofLutProvider.FILE,
+    # tof_lut_provider: TofLutProvider = TofLutProvider.FILE,
 ) -> sciline.Pipeline:
     """
     Generic workflow for computing the neutron time-of-flight for detector and monitor
@@ -75,16 +78,16 @@ def GenericTofWorkflow(
     for provider in eto_to_tof.providers():
         wf.insert(provider)
 
-    if tof_lut_provider == TofLutProvider.FILE:
-        wf.insert(load_tof_lookup_table)
-    else:
-        wf.insert(eto_to_tof.compute_tof_lookup_table)
-        if tof_lut_provider == TofLutProvider.TOF:
-            wf.insert(simulation.simulate_chopper_cascade_using_tof)
-        if tof_lut_provider == TofLutProvider.MCSTAS:
-            raise NotImplementedError("McStas simulation not implemented yet")
+    # if tof_lut_provider == TofLutProvider.FILE:
+    wf.insert(load_tof_lookup_table)
+    # else:
+    #     wf.insert(eto_to_tof.compute_tof_lookup_table)
+    #     if tof_lut_provider == TofLutProvider.TOF:
+    #         wf.insert(simulation.simulate_chopper_cascade_using_tof)
+    #     if tof_lut_provider == TofLutProvider.MCSTAS:
+    #         raise NotImplementedError("McStas simulation not implemented yet")
 
-    for key, value in eto_to_tof.default_parameters().items():
-        wf[key] = value
+    # Default parameters
+    wf[PulseStrideOffset] = None
 
     return wf
